Add check for available memory
Using these playbooks to install DeskPRO on a server requires 512-1024MB
of memory. Running with less than that may cause different systems to
fail, such as having MySQL or Elasticsearch be killed due to not enough
memory available.

This is not a supported scenario and failures are quite likely to
happen.