Collect all useful output into a log file
The main script may omit some output for the sake of being user
friendly, but sometimes things fail, and in those cases we need as much
information as possible to troubleshoot the issue.

This saves all output into a file identified by `FULL_LOG_FILE` which
can be set on the command line using the `--log` option, and defaults to
a random file on `/tmp`.

If the install fails for some reason, the logs are uploaded to DeskPRO,
which will aid in debugging and improving the scripts.