Fix installation on Debian systems
Debian has, by default, a rather old version of Ansible (1.7.2). It's
still recent enough for our purposes, but some Ansible constructs that
were being used in newer roles were preventing the install to finish.

In this commmit we fix the versions of roles to versions that are known
to work in all the required environments with the pre-packaged version
of Ansible so we don't have to build it ourselves.

Also, some base packages were missing from Debian on EC2, so they were
pre-installed as well.