Back to the basics by installing more lower level tools
Some tools aren't installed by default on more "restricted" platforms,
such as on the official LXC template for Debian. This accounts for those
tools.

Also, due to updated packages present on jessie-backports,
python-pkg-resources is now installed manually. This package will be
installed at the same time as Ansible if not present, and there's a new
version on jessie-backports that conflicts with python-pip. By
installing it here, the playbook itself stays the same while we avoid a
"dependency race" as the Ansible package from jessie-backports will
happily use the older python-pkg-resources

One small fix: skip update_cache due to ansible 2.2 bug