Update platforms under test
Ubuntu 18.04 was recently released and was added.

We're still testing against Debian 8, but `debian-8.6` was using an old
kernel that was starting to have issues with recent iptables updates.
Switching to the "base" `debian-8` means I don't have to update this
list as often while still testing against the most likely scenario for
new deployments.