Install the Azul Java distribution on Debian Jessie
Jessie is in LTS mode with a little less than a year of support, so
usually people shouldn't be installing a new system into it. But this
makes it possible to install Java 8 (which is required by Elasticsearch)
into Jessie. The previous installer version used the backports
repository, but that's been discontinued and even the version on
archive.debian.org is not valid anymore so we needed an alternative here